+++ src/jsx/components/friends.jsx
