+++ src/jsx/dispatcher/dispatcher.jsx
@@ -1,0 +1,3 @@
+var Dispatcher = require('flux').Dispatcher;
+
+module.exports = new Dispatcher();
